Move QUICServer to lib/proxy/peer/quic
